requirements.txt
playwright>=1.45.0
pydantic>=2.7.0
pyyaml>=6.0.1
tenacity>=8.5.0
loguru>=0.7.2
pandas>=2.2.0
openpyxl>=3.1.2
python-dotenv>=1.0.1
